requirements.txt
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
python-dotenv
drf-yasg
requests